requirements.txt
mujoco-py<2.2,>=2.1
mujoco
cython<3
numpy
pyquaternion
termcolor
numpy-stl
trimesh
fast_simplification